add __eq__ TensorBlock and TensorMap
first attempt at the __eq__ function for TensorBlock/TensorMap
it is wrong for component > 0
but so far even if the other label are shuffled it can reconstract the equality.
for example it understand that if two block have sample in different order they can be equal.
It also understatnd if labels have the names in different order.
the __eq__ for labels is not implemented becouse it would need the `position` function that now works only if the labels comes from a TensorMap/Block